move code to useCustomHook
